requirements.txt
black
ruff

pyyaml
tabulate